+++ packages/headless/src/controllers/controller/headless-controller.ts
@@ -2,10 +2,10 @@ import {Engine} from '../../app/headless-engine';
 
 export type Controller = ReturnType<typeof buildController>;
 
-export function buildController(engine: Engine) {
+export function buildController<T>(engine: Engine<T>) {
   let prevState = '{}';
 
-  const hasStateChanged = (currentState: Record<string, any>): boolean => {
+  const hasStateChanged = (currentState: T): boolean => {
     try {
       const stringifiedState = JSON.stringify(currentState);
       const hasChanged = prevState !== stringifiedState;
@@ -37,7 +37,7 @@ export function buildController(engine: Engine) {
     },
 
     get state() {
-      return {};
+      return {} as T;
     },
   };
 }