+++ packages/headless/src/test/mock-advanced-search-parameters-state.ts
@@ -1,0 +1,11 @@
+import {AdvancedSearchQueriesState} from '../features/advanced-search-queries/advanced-search-queries-slice';
+
+export function buildMockAdvancedSearchQueriesState(
+  config: Partial<AdvancedSearchQueriesState> = {}
+): AdvancedSearchQueriesState {
+  return {
+    aq: '',
+    cq: '',
+    ...config,
+  };
+}