+++ packages/headless/src/test/mock-state.ts
@@ -1,4 +1,3 @@
-import {SearchPageState} from '../state';
 import {getConfigurationInitialState} from '../features/configuration/configuration-slice';
 import {getQueryInitialState} from '../features/query/query-slice';
 import {getRedirectionInitialState} from '../features/redirection/redirection-slice';
@@ -17,18 +16,17 @@ import {getPipelineInitialState} from '../features/pipeline/pipeline-slice';
 import {makeHistory} from '../app/undoable';
 import {getSearchHubInitialState} from '../features/search-hub/search-hub-slice';
 import {getCategoryFacetSetInitialState} from '../features/facets/category-facet-set/category-facet-set-slice';
-import {getInitialConstantQueryState} from '../features/constant-query/constant-query-slice';
 import {getFacetSearchSetInitialState} from '../features/facets/facet-search-set/specific/specific-facet-search-set-slice';
 import {getCategoryFacetSearchSetInitialState} from '../features/facets/facet-search-set/category/category-facet-search-set-slice';
-import {getInitialAdvancedQueryState} from '../features/advanced-query/advanced-query-slice';
+import {getAdvancedSearchQueriesInitialState} from '../features/advanced-search-queries/advanced-search-queries-slice';
+import {SearchAppState} from '../state/search-app-state';
 
 export function createMockState(
-  config: Partial<SearchPageState> = {}
-): SearchPageState {
+  config: Partial<SearchAppState> = {}
+): SearchAppState {
   return {
     configuration: getConfigurationInitialState(),
-    constantQuery: getInitialConstantQueryState(),
-    advancedQuery: getInitialAdvancedQueryState(),
+    advancedSearchQueries: getAdvancedSearchQueriesInitialState(),
     facetSet: getFacetSetInitialState(),
     dateFacetSet: getDateFacetSetInitialState(),
     numericFacetSet: getNumericFacetSetInitialState(),