+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.8](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.7...v0.1.0-alpha.8) (2020-10-20)
+
+
+### Features
+
+* **headless:** basic recommendation state + reducers ([51717fd](https://bitbucket.org/coveord/ui-kit/commits/51717fd006d147a029a8c588ea6e5ee0c5400bce))
+
+
+
+
+
+# [0.1.0-alpha.7](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.6...v0.1.0-alpha.7) (2020-10-20)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [0.1.0-alpha.6](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.5...v0.1.0-alpha.6) (2020-10-19)
+
+
+### Features
+
+* **headless:** Refactor cq and aq as string in an single object ([7b0010a](https://bitbucket.org/coveord/ui-kit/commits/7b0010a4b75f8e1faef28a9211cc6596cc8f3bd8))
+
+
+
+
+
 # 0.1.0-alpha.5 (2020-10-16)